+++ components/module/example/demo-input.tsx
@@ -1,12 +1,20 @@
 "use client";
 import CButton from "@/components/shared/custome/c-button";
 import CInput from "@/components/shared/custome/c-input";
-import FieldSelect from "@/components/shared/field-select";
+import CInputField from "@/components/shared/custome/c-input-field";
+import CSelect from "@/components/shared/custome/c-select";
+import CSelectField from "@/components/shared/custome/c-select-field";
 import { OPT_DUMMY } from "@/lib/constants";
 import { UserSchema } from "@/lib/validation";
 import { useForm } from "@tanstack/react-form";
+import { useState } from "react";
 
 export default function DemoInput() {
+  const [payload, setPayload] = useState({
+    search: "",
+    type: "",
+  });
+
   const form = useForm({
     defaultValues: {
       name: "",
@@ -33,7 +41,7 @@ export default function DemoInput() {
   };
 
   return (
-    <div className="size-full max-w-2xl mx-auto flex items-center justify-center ">
+    <div className="size-full max-w-2xl mx-auto flex items-start justify-center gap-[8px]">
       <form
         className="size-full"
         onSubmit={(e) => {
@@ -54,7 +62,7 @@ export default function DemoInput() {
         />
         <CInput name="email" placeholder="example@gmail.com" form={form} />
         <CInput name="password" placeholder="password" form={form} />
-        <FieldSelect
+        <CSelect
           name="hobby"
           placeholder="Hobby"
           options={OPT_DUMMY}
@@ -62,6 +70,20 @@ export default function DemoInput() {
         />
         <CButton title="Submit" type="submit" />
       </form>
+
+      <div className="w-full grid grid-cols-1 gap-[8px]">
+        <CInputField
+          placeholder="ajsja"
+          value={payload.search}
+          onChange={(e) => setPayload({ ...payload, search: e.target.value })}
+        />
+        <CSelectField
+          placeholder="select type"
+          options={OPT_DUMMY}
+          value={payload.type}
+          onChange={(e) => setPayload({ ...payload, type: e })}
+        />
+      </div>
     </div>
   );
 }